requirements.txt
tkinter
requests
pillow
pygame
